Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -2,7 +2,7 @@
 int main() {
     int a,b;
     char operators;
-    scanf("%d %d %c",&a,&b,&operators);
+    scanf("%d %d %c", &a ,&b , &operators);
     if (operators == '+') {
         printf("%d",a+b);
     }
@@ -13,7 +13,7 @@ int main() {
         printf("%d",a*b);
     }
     else if (operators == '/') {
-        printf("%f",a/b);
+        printf("%f",(float)a/b);
     }
     else {
         printf("Yoyo");
